requirements.txt
Flask==2.3.2
gunicorn==20.1.0
torch>=2.0.0
transformers>=4.30.0
keybert>=0.7.0
tweepy>=4.14.0
python-linkedin-v2>=0.0.6
python-dotenv>=1.0.0
Pillow>=9.5.0
keybert>=0.7.0